Enhanced logging middleware to display session info and edited authentication middleware to create a cart for every new session
