+++ server/app/configure/logging.js
@@ -1,0 +1,17 @@
+'use strict';
+var chalk = require('chalk');
+var util = require('util');
+
+
+var logMiddleware = function (req, res, next) {
+    util.log(('---NEW REQUEST---'));
+    console.log(util.format(chalk.green('%s: %s'), 'SESSION ', util.inspect(req.session)));
+    console.log(util.format(chalk.red('%s: %s %s'), 'REQUEST ', req.method, req.path));
+    console.log(util.format(chalk.yellow('%s: %s'), 'QUERY   ', util.inspect(req.query)));
+    console.log(util.format(chalk.cyan('%s: %s'), 'BODY    ', util.inspect(req.body)));
+    next();
+};
+
+module.exports = function (app) {
+    app.setValue('log', logMiddleware);
+};
